refactor import paths for MCP utilities to improve module organization
